+++ packages/@baldeweg/ui/src/composables.js
@@ -1,3 +1,2 @@
 export * from './composables/useToast.js'
-export * from './composables/useLocale.js'
 export * from './composables/useRequest.js'